[Flow][NFC] Move cloning/moving helpers to RegionOpUtils
This makes these helper functions usable from transform dialect op
implementations and the Flow/Transforms.
